requirements-phase0.txt
requests>=2.31.0
websockets>=12.0
python-dotenv>=1.0.0
psycopg[binary]>=3.1.0
tenacity>=8.2.0
